+++ app/index.js
@@ -1,5 +1,4 @@
 const express = require('express');
-const { Pool } = require('pg');
 const path = require('path');
 
 const app = express();
@@ -9,35 +8,16 @@ const PORT = 3000;
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 
-// PostgreSQL
-const pool = new Pool({
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_NAME,
-});
-
+// Import all routers
+const products = require('./router/products');
 
 // Endpoints
 app.get('/', (req, res) => {
     res.sendFile(path.join(__dirname, 'frontend', 'index.html'));
 });
 
-
-// === Externe Endpoints und Connectors === || === To Do
-app.get('/api/products', async (req, res) => {
-    try {
-        const result = await pool.query('SELECT * FROM webshop.products');
-        const data = JSON.stringify(result.rows);
-
-        res.writeHead(200, { 'Content-Type': 'application/json' });
-        res.end(data);
-    } catch (err) {
-        // === Genauere Fehler Meldungen === || === To Do
-        res.writeHead(500, { 'Content-Type': 'text/plain' });
-        res.end('Interner Server Fehler');
-    }
-});
+// Use the imported Routers
+app.use('/api/products', products);
 
 // Server starten
 app.listen(PORT, () => {
